Modify the sessionId param as int
:Release Notes:
SessionId parameter of APIs are modified as integer

:Detailed Notes:
As this service is having OSE layer and in OSE
sessionmanager is not present.
Session Id's in OSE are int type.Hence to align the code base
to OSE changing sessionId parameter to int.

:Testing Performed:
Tested on webos build rebased on #2026 official binary
Test case verified: QWA-7483, QWA-7449
Test results are attached to PLAT-131534

:QA Notes:
NA

:Issues Addressed:
[PLAT-131534] SessionId to be changed to integer from string in APIs

Change-Id: I695748d8f5aafbce22111ce52d9d5db1e484caba
Reviewed-on: http://gpro.lge.com/c/webos-pro/com.webos.service.audiofocusmanager/+/300898
Tested-by: <thomas.georgec@lge.com>
Reviewed-by: <thomas.georgec@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Velicheti Sita <velicheti.sita@lge.com>
Reviewed-by: Sushovan Gangopadhy <sushovan.g@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.audiofocusmanager/+/303644
Tested-by: aleti.nageshwar <aleti.nageshwar@lge.com>
Reviewed-by: aleti.nageshwar <aleti.nageshwar@lge.com>